formatted

+++ forge_contract/test/CrowdFundingTest.t.sol
@@ -28,38 +28,28 @@ contract CrowdFundingTest is Test {
     uint256 constant SEND_VALUE = 0.1 ether;
     DeployCrowdFundingWithCampaigns deployCrowdFunding;
 
-    function getTestCampaignInput()
-        internal
-        view
-        returns (CampaignLib.CampaignInput memory)
-    {
-        return
-            CampaignLib.CampaignInput({
-                title: "Test Campaign",
-                description: "This is a test campaign",
-                target: 100,
-                deadline: block.timestamp + 1000,
-                image: "https://example.com/image.png",
-                category: CampaignLib.Category.Technology,
-                allowFlexibleWithdrawal: false
-            });
+    function getTestCampaignInput() internal view returns (CampaignLib.CampaignInput memory) {
+        return CampaignLib.CampaignInput({
+            title: "Test Campaign",
+            description: "This is a test campaign",
+            target: 100,
+            deadline: block.timestamp + 1000,
+            image: "https://example.com/image.png",
+            category: CampaignLib.Category.Technology,
+            allowFlexibleWithdrawal: false
+        });
     }
 
-    function getTestCampaignInputFlexibleWithdrawal()
-        internal
-        view
-        returns (CampaignLib.CampaignInput memory)
-    {
-        return
-            CampaignLib.CampaignInput({
-                title: "Test Campaign Flexible",
-                description: "This is a test campaign with flexible withdrawal",
-                target: 100,
-                deadline: block.timestamp + 1000,
-                image: "https://example.com/image.png",
-                category: CampaignLib.Category.Technology,
-                allowFlexibleWithdrawal: true
-            });
+    function getTestCampaignInputFlexibleWithdrawal() internal view returns (CampaignLib.CampaignInput memory) {
+        return CampaignLib.CampaignInput({
+            title: "Test Campaign Flexible",
+            description: "This is a test campaign with flexible withdrawal",
+            target: 100,
+            deadline: block.timestamp + 1000,
+            image: "https://example.com/image.png",
+            category: CampaignLib.Category.Technology,
+            allowFlexibleWithdrawal: true
+        });
     }
 
     function getTestCampaignUpdateInput(uint256 campaignId)
@@ -67,16 +57,15 @@ contract CrowdFundingTest is Test {
         view
         returns (CampaignLib.CampaignUpdateInput memory)
     {
-        return
-            CampaignLib.CampaignUpdateInput({
-                id: campaignId,
-                title: "Updated Test Campaign",
-                description: "This is an updated test campaign",
-                target: 200,
-                deadline: block.timestamp + 2000,
-                image: "https://example.com/updated-image.png",
-                category: CampaignLib.Category.Art
-            });
+        return CampaignLib.CampaignUpdateInput({
+            id: campaignId,
+            title: "Updated Test Campaign",
+            description: "This is an updated test campaign",
+            target: 200,
+            deadline: block.timestamp + 2000,
+            image: "https://example.com/updated-image.png",
+            category: CampaignLib.Category.Art
+        });
     }
 
     function setUp() public {
@@ -104,15 +93,9 @@ contract CrowdFundingTest is Test {
         assertEq(crowdFunding.campaignCount(), campaignCount + 1);
 
         // verify that the campaign is in draft status
-        assertEq(
-            uint8(crowdFunding.getCampaignStatus(campaignCount)),
-            uint8(CampaignLib.CampaignStatus.Draft)
-        );
+        assertEq(uint8(crowdFunding.getCampaignStatus(campaignCount)), uint8(CampaignLib.CampaignStatus.Draft));
 
-        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(
-            campaignCount,
-            false
-        );
+        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(campaignCount, false);
         assertEq(campaign.owner, USER);
         assertEq(campaign.title, "Test Campaign");
         assertEq(campaign.target, 100);
@@ -128,15 +111,9 @@ contract CrowdFundingTest is Test {
         assertEq(crowdFunding.campaignCount(), campaignCount + 1);
 
         // verify that the campaign is in published status
-        assertEq(
-            uint8(crowdFunding.getCampaignStatus(campaignCount)),
-            uint8(CampaignLib.CampaignStatus.Published)
-        );
+        assertEq(uint8(crowdFunding.getCampaignStatus(campaignCount)), uint8(CampaignLib.CampaignStatus.Published));
 
-        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(
-            campaignCount,
-            true
-        );
+        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(campaignCount, true);
         assertEq(campaign.owner, USER);
         assertEq(campaign.title, "Test Campaign");
         assertEq(campaign.target, 100);
@@ -148,10 +125,7 @@ contract CrowdFundingTest is Test {
         uint256 campaignCount = crowdFunding.campaignCount();
         crowdFunding.createDraft(getTestCampaignInput());
 
-        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(
-            campaignCount,
-            false
-        );
+        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(campaignCount, false);
         assertEq(campaign.owner, USER);
         assertEq(campaign.title, "Test Campaign");
 
@@ -174,10 +148,7 @@ contract CrowdFundingTest is Test {
         crowdFunding.publishCampaign(campaignCount);
 
         // verify that the campaign is published
-        assertEq(
-            uint8(crowdFunding.getCampaignStatus(campaignCount)),
-            uint8(CampaignLib.CampaignStatus.Published)
-        );
+        assertEq(uint8(crowdFunding.getCampaignStatus(campaignCount)), uint8(CampaignLib.CampaignStatus.Published));
 
         vm.stopPrank();
     }
@@ -190,10 +161,7 @@ contract CrowdFundingTest is Test {
         // Update the draft campaign
         crowdFunding.updateDraftCampaign(getTestCampaignUpdateInput(campaignCount));
 
-        CampaignLib.Campaign memory updatedCampaign = crowdFunding.getCampaign(
-            campaignCount,
-            false
-        );
+        CampaignLib.Campaign memory updatedCampaign = crowdFunding.getCampaign(campaignCount, false);
 
         // Verify the updates
         assertEq(updatedCampaign.title, "Updated Test Campaign");
@@ -213,10 +181,7 @@ contract CrowdFundingTest is Test {
         crowdFunding.cancelDraftCampaign(campaignCount);
 
         // Verify that the campaign is cancelled
-        assertEq(
-            uint8(crowdFunding.getCampaignStatus(campaignCount)),
-            uint8(CampaignLib.CampaignStatus.Cancelled)
-        );
+        assertEq(uint8(crowdFunding.getCampaignStatus(campaignCount)), uint8(CampaignLib.CampaignStatus.Cancelled));
 
         vm.stopPrank();
     }
@@ -228,10 +193,7 @@ contract CrowdFundingTest is Test {
 
         crowdFunding.donateToCampaign{value: 10}(campaignCount);
 
-        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(
-            campaignCount,
-            true
-        );
+        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(campaignCount, true);
 
         // verify that the campaign has received the donation
         assertEq(campaign.amountCollected, 10);
@@ -248,8 +210,7 @@ contract CrowdFundingTest is Test {
         crowdFunding.createPublishedCampaign(getTestCampaignInput());
 
         crowdFunding.donateToCampaign{value: 10}(campaignCount);
-        (uint256 amount, uint256 timestamp, uint16 noOfDonations) = crowdFunding
-            .getDonorInfo(campaignCount, USER);
+        (uint256 amount, uint256 timestamp, uint16 noOfDonations) = crowdFunding.getDonorInfo(campaignCount, USER);
         assertEq(amount, 10);
         assertEq(timestamp, block.timestamp);
         assertEq(noOfDonations, 1);
@@ -263,8 +224,7 @@ contract CrowdFundingTest is Test {
 
         crowdFunding.donateToCampaign{value: 10}(campaignCount);
         crowdFunding.donateToCampaign{value: 20}(campaignCount);
-        (uint256 amount, uint256 timestamp, uint16 noOfDonations) = crowdFunding
-            .getDonorInfo(campaignCount, USER);
+        (uint256 amount, uint256 timestamp, uint16 noOfDonations) = crowdFunding.getDonorInfo(campaignCount, USER);
         assertEq(amount, 30);
         assertEq(timestamp, block.timestamp);
         assertEq(noOfDonations, 2);
@@ -296,9 +256,7 @@ contract CrowdFundingTest is Test {
         // 1. USER creates a published campaign with flexible withdrawal
         vm.startPrank(USER);
         uint256 campaignCount = crowdFunding.campaignCount();
-        crowdFunding.createPublishedCampaign(
-            getTestCampaignInputFlexibleWithdrawal()
-        );
+        crowdFunding.createPublishedCampaign(getTestCampaignInputFlexibleWithdrawal());
         vm.stopPrank();
 
         // 2. ALICE donates to the campaign
@@ -322,10 +280,7 @@ contract CrowdFundingTest is Test {
         assertEq(balanceAfter, balanceBefore + SEND_VALUE);
 
         // 7. Check that the campaign state is updated correctly
-        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(
-            campaignCount,
-            true
-        );
+        CampaignLib.Campaign memory campaign = crowdFunding.getCampaign(campaignCount, true);
         assertEq(campaign.amountCollected, SEND_VALUE);
         assertEq(campaign.withdrawnAmount, SEND_VALUE);
         assertEq(crowdFunding.getRemainingBalance(campaignCount), 0);
@@ -334,42 +289,42 @@ contract CrowdFundingTest is Test {
     function test_getCampaigns() public {
         vm.startPrank(USER);
         uint256 campaignCount = crowdFunding.campaignCount();
-        
+
         // Create multiple campaigns
         crowdFunding.createDraft(getTestCampaignInput());
         crowdFunding.createPublishedCampaign(getTestCampaignInput());
-        
+
         // Get all campaigns
         CampaignLib.Campaign[] memory allCampaigns = crowdFunding.getCampaigns();
         assertEq(allCampaigns.length, campaignCount + 2);
-        
+
         // Verify campaign details
         // assertEq(allCampaigns[0].owner, USER);
         // assertEq(allCampaigns[1].owner, USER);
         // assertEq(uint8(allCampaigns[0].status), uint8(CampaignLib.CampaignStatus.Draft));
         // assertEq(uint8(allCampaigns[1].status), uint8(CampaignLib.CampaignStatus.Published));
-        
+
         vm.stopPrank();
     }
 
     function test_getPublishedCampaigns() public {
         vm.startPrank(USER);
         uint256 currentPublishedCampaigns = crowdFunding.getPublishedCampaigns2().length;
-        
+
         // Create mix of draft and published campaigns
         crowdFunding.createDraft(getTestCampaignInput());
         crowdFunding.createPublishedCampaign(getTestCampaignInput());
         crowdFunding.createPublishedCampaign(getTestCampaignInputFlexibleWithdrawal());
-        
+
         // Get only published campaigns
         CampaignLib.Campaign[] memory publishedCampaigns = crowdFunding.getPublishedCampaigns2();
         assertEq(publishedCampaigns.length, currentPublishedCampaigns + 2);
-        
+
         // All returned campaigns should be published
-        for (uint i = 0; i < publishedCampaigns.length; i++) {
+        for (uint256 i = 0; i < publishedCampaigns.length; i++) {
             assertEq(uint8(publishedCampaigns[i].status), uint8(CampaignLib.CampaignStatus.Published));
         }
-        
+
         vm.stopPrank();
     }
 
@@ -386,8 +341,8 @@ contract CrowdFundingTest is Test {
         // Get USER's campaigns
         CampaignLib.Campaign[] memory userCampaigns = crowdFunding.getUserCampaigns(USER);
         assertEq(userCampaigns.length, 2);
-        
-        for (uint i = 0; i < userCampaigns.length; i++) {
+
+        for (uint256 i = 0; i < userCampaigns.length; i++) {
             assertEq(userCampaigns[i].owner, USER);
         }
 
@@ -399,7 +354,7 @@ contract CrowdFundingTest is Test {
 
     function test_getCampaignsByCategory() public {
         vm.startPrank(USER);
-        
+
         // Create campaigns with different categories
         CampaignLib.CampaignInput memory techInput = getTestCampaignInput();
         techInput.category = CampaignLib.Category.Technology;
@@ -412,14 +367,15 @@ contract CrowdFundingTest is Test {
         CampaignLib.CampaignInput memory techInput2 = getTestCampaignInput();
         techInput2.category = CampaignLib.Category.Technology;
         crowdFunding.createPublishedCampaign(techInput2);
-        
+
         vm.stopPrank();
 
         // Get campaigns by Technology category
-        CampaignLib.Campaign[] memory techCampaigns = crowdFunding.getCampaignsByCategory(CampaignLib.Category.Technology);
+        CampaignLib.Campaign[] memory techCampaigns =
+            crowdFunding.getCampaignsByCategory(CampaignLib.Category.Technology);
         assertEq(techCampaigns.length, 2);
-        
-        for (uint i = 0; i < techCampaigns.length; i++) {
+
+        for (uint256 i = 0; i < techCampaigns.length; i++) {
             assertEq(uint8(techCampaigns[i].category), uint8(CampaignLib.Category.Technology));
             assertEq(uint8(techCampaigns[i].status), uint8(CampaignLib.CampaignStatus.Published));
         }
@@ -435,7 +391,7 @@ contract CrowdFundingTest is Test {
         vm.startPrank(USER);
         CampaignLib.CampaignInput memory input = getTestCampaignInput();
         input.target = 0; // Invalid target
-        
+
         vm.expectRevert(CrowdFunding.InvalidTarget.selector);
         crowdFunding.createDraft(input);
         vm.stopPrank();
@@ -445,7 +401,7 @@ contract CrowdFundingTest is Test {
         vm.startPrank(USER);
         CampaignLib.CampaignInput memory input = getTestCampaignInput();
         input.deadline = block.timestamp - 1; // Past deadline
-        
+
         vm.expectRevert(CrowdFunding.InvalidDeadline.selector);
         crowdFunding.createDraft(input);
         vm.stopPrank();
@@ -454,7 +410,7 @@ contract CrowdFundingTest is Test {
     function test_donateToCampaign_ZeroDonation() public {
         vm.startPrank(USER);
         uint256 campaignId = crowdFunding.createPublishedCampaign(getTestCampaignInput());
-        
+
         vm.expectRevert(CrowdFunding.ZeroDonation.selector);
         crowdFunding.donateToCampaign{value: 0}(campaignId);
         vm.stopPrank();
